Collect symbol builds; restructure rust directories
