+++ bindings/rust/src/lib.rs
