--- rust/rust_wrapper.c
@@ -1,28 +1,0 @@
-/* Copyright (c) 2022, Google Inc.
- *
- * Permission to use, copy, modify, and/or distribute this software for any
- * purpose with or without fee is hereby granted, provided that the above
- * copyright notice and this permission notice appear in all copies.
- *
- * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
- * WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
- * MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY
- * SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
- * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION
- * OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN
- * CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE. */
-
-#include "rust_wrapper.h"
-
-
-int ERR_GET_LIB_RUST(uint32_t packed_error) {
-  return ERR_GET_LIB(packed_error);
-}
-
-int ERR_GET_REASON_RUST(uint32_t packed_error) {
-  return ERR_GET_REASON(packed_error);
-}
-
-int ERR_GET_FUNC_RUST(uint32_t packed_error) {
-  return ERR_GET_FUNC(packed_error);
-}
