--- rust/wrapper.h
@@ -1,78 +1,0 @@
-#include "../include/openssl/aes.h"
-#include "../include/openssl/asn1.h"
-#include "../include/openssl/asn1_mac.h"
-#include "../include/openssl/asn1t.h"
-#include "../include/openssl/base.h"
-#include "../include/openssl/base64.h"
-#include "../include/openssl/bio.h"
-#include "../include/openssl/blake2.h"
-#include "../include/openssl/blowfish.h"
-#include "../include/openssl/bn.h"
-#include "../include/openssl/buf.h"
-#include "../include/openssl/buffer.h"
-#include "../include/openssl/bytestring.h"
-#include "../include/openssl/cast.h"
-#include "../include/openssl/chacha.h"
-#include "../include/openssl/cipher.h"
-#include "../include/openssl/cmac.h"
-#include "../include/openssl/conf.h"
-#include "../include/openssl/cpu.h"
-#include "../include/openssl/crypto.h"
-#include "../include/openssl/curve25519.h"
-#include "../include/openssl/des.h"
-#include "../include/openssl/dh.h"
-#include "../include/openssl/digest.h"
-#include "../include/openssl/dsa.h"
-#include "../include/openssl/dtls1.h"
-#include "../include/openssl/e_os2.h"
-#include "../include/openssl/ec.h"
-#include "../include/openssl/ec_key.h"
-#include "../include/openssl/ecdh.h"
-#include "../include/openssl/ecdsa.h"
-#include "../include/openssl/engine.h"
-#include "../include/openssl/err.h"
-#include "../include/openssl/evp.h"
-#include "../include/openssl/evp_errors.h"
-#include "../include/openssl/ex_data.h"
-#include "../include/openssl/hkdf.h"
-#include "../include/openssl/hmac.h"
-#include "../include/openssl/hpke.h"
-#include "../include/openssl/hrss.h"
-#include "../include/openssl/is_boringssl.h"
-#include "../include/openssl/lhash.h"
-#include "../include/openssl/md4.h"
-#include "../include/openssl/md5.h"
-#include "../include/openssl/mem.h"
-#include "../include/openssl/obj.h"
-#include "../include/openssl/obj_mac.h"
-#include "../include/openssl/objects.h"
-#include "../include/openssl/opensslconf.h"
-#include "../include/openssl/opensslv.h"
-#include "../include/openssl/ossl_typ.h"
-#include "../include/openssl/pem.h"
-#include "../include/openssl/pkcs12.h"
-#include "../include/openssl/pkcs7.h"
-#include "../include/openssl/pkcs8.h"
-#include "../include/openssl/poly1305.h"
-#include "../include/openssl/pool.h"
-#include "../include/openssl/rand.h"
-#include "../include/openssl/rc4.h"
-#include "../include/openssl/ripemd.h"
-#include "../include/openssl/rsa.h"
-#include "../include/openssl/safestack.h"
-#include "../include/openssl/sha.h"
-#include "../include/openssl/siphash.h"
-#include "../include/openssl/span.h"
-#include "../include/openssl/srtp.h"
-#include "../include/openssl/ssl.h"
-#include "../include/openssl/ssl3.h"
-#include "../include/openssl/stack.h"
-#include "../include/openssl/thread.h"
-#include "../include/openssl/tls1.h"
-#include "../include/openssl/trust_token.h"
-#include "../include/openssl/type_check.h"
-#include "../include/openssl/x509.h"
-#include "../include/openssl/x509_vfy.h"
-#include "../include/openssl/x509v3.h"
-
-#include "rust_wrapper.h"
